BUG: Fixed build errors on linux.

+++ minerva/minerva/classifiers/test/test-stacked-autoencoder.cpp
@@ -78,7 +78,7 @@ static Image generateRandomImage(size_t xPixels, size_t yPixels, size_t colors,
 	{
 		for(size_t x = 0; x != xPixels; ++x)
 		{
-			for(int c = 0; c != colors; ++c)
+			for(size_t c = 0; c != colors; ++c)
 			{
 				float value = distribution(engine);
 
@@ -105,7 +105,7 @@ static Image addRandomNoiseToImage(const Image& image, float noiseMagnitude,
 	{
 		for(size_t x = 0; x != xPixels; ++x)
 		{
-			for(int c = 0; c != colors; ++c)
+			for(size_t c = 0; c != colors; ++c)
 			{
 				float value = distribution(engine) + image.getStandardizedComponentAt(x, y, c);
